# src/main/java/com/greendelta/collaboration/controller/user/ReviewController.java
package com.greendelta.collaboration.controller.user;

import java.util.Collections;
import java.util.List;
import java.util.Map;
import java.util.stream.Collectors;

import org.openlca.git.model.Reference;
import org.openlca.util.Strings;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.greendelta.collaboration.controller.util.FrontendReference;
import com.greendelta.collaboration.controller.util.Response;
import com.greendelta.collaboration.controller.util.Reviews;
import com.greendelta.collaboration.error.WebRequestException;
import com.greendelta.collaboration.model.settings.ServerSetting;
import com.greendelta.collaboration.model.task.Review;
import com.greendelta.collaboration.model.task.ReviewReference;
import com.greendelta.collaboration.model.task.TaskState;
import com.greendelta.collaboration.service.Repository;
import com.greendelta.collaboration.service.RepositoryService;
import com.greendelta.collaboration.service.SettingsService;
import com.greendelta.collaboration.service.task.ReviewService;
import com.greendelta.collaboration.service.task.TaskService;
import com.greendelta.collaboration.service.user.AccessService;
import com.greendelta.collaboration.service.user.NotificationService;
import com.greendelta.collaboration.service.user.UserService;

@RestController
@RequestMapping("ws/task/review")
public class ReviewController {

	private final ReviewService service;
	private final TaskService taskService;
	private final UserService userService;
	private final AccessService accessService;
	private final NotificationService notificationService;
	private final RepositoryService repoService;
	private final SettingsService settingsService;

	@Autowired
	public ReviewController(ReviewService service, TaskService taskService, UserService userService,
			AccessService accessService, NotificationService notificationService, RepositoryService repoService,
			SettingsService settingsService) {
		this.service = service;
		this.taskService = taskService;
		this.userService = userService;
		this.accessService = accessService;
		this.notificationService = notificationService;
		this.repoService = repoService;
		this.settingsService = settingsService;
	}

	@GetMapping("{id}")
	public Map<String, Object> get(@PathVariable("id") long id) {
		if (!settingsService.is(ServerSetting.TASKS_ENABLED))
			throw Response.unavailable("Task feature not enabled");
		var review = service.get(id);
		if (review == null)
			throw Response.notFound("No review with id " + id + " found");
		try (var repo = repoService.get(review.repositoryPath)) {
			return Reviews.map(review, repo);
		}
	}

	@PostMapping
	public Map<String, Object> start(@RequestBody Review review) {
		if (!settingsService.is(ServerSetting.TASKS_ENABLED))
			throw Response.unavailable("Task feature not enabled");
		checkValidity(review);
		try (var repo = repoService.get(review.repositoryPath)) {
			service.start(review);
			notificationService.taskStarted(repo, review).send();
			return getActiveTasks();
		}
	}

	@PutMapping("{id}")
	public Map<String, Object> update(@RequestBody Review review) {
		if (!settingsService.is(ServerSetting.TASKS_ENABLED))
			throw Response.unavailable("Task feature not enabled");
		if (Strings.nullOrEmpty(review.name))
			throw Response.badRequest("name", "Missing input: Name");
		service.merge(review);
		return getActiveTasks();
	}

	private void checkValidity(Review review) {
		if (Strings.nullOrEmpty(review.name))
			throw Response.badRequest("name", "Missing input: Name");
		if (Strings.nullOrEmpty(review.repositoryPath))
			throw Response.badRequest("repositoryPath", "Missing input: Repository path");
		if (!review.assignments.isEmpty() || review.id == 0)
			throw Response.conflict("Review object already exists");
	}

	@PutMapping("{id}/references")
	public Map<String, Object> setReferences(
			@PathVariable("id") long id,
			@RequestBody List<FrontendReference> references) {
		if (!settingsService.is(ServerSetting.TASKS_ENABLED))
			throw Response.unavailable("Task feature not enabled");
		var review = service.get(id);
		if (review == null)
			throw Response.notFound("No review with id " + id + " found");
		if (!review.assignments.isEmpty())
			throw Response.badRequest("", "References can not be changed after reviewer were already assigned");
		try (var repo = repoService.get(review.repositoryPath)) {
			if (repo == null)
				throw Response.notFound("No repository with id " + review.repositoryPath + " found");
			var refs = FrontendReference.collect(repo, references);
			service.setReferences(id, refs.stream().map(ref -> convert(repo, ref)).collect(Collectors.toSet()));
			return getActiveTasks();
		}
	}

	@PutMapping("{id}/complete")
	public Map<String, Object> completeReview(@PathVariable("id") long id) {
		if (!settingsService.is(ServerSetting.TASKS_ENABLED))
			throw Response.unavailable("Task feature not enabled");
		Review review = service.get(id);
		if (review == null)
			throw Response.notFound("No review with id " + id + " found");
		try (var repo = repoService.get(review.repositoryPath)) {
			service.end(review, TaskState.COMPLETED);
			notificationService.taskCompleted(repo, review).send();
			return getActiveTasks();
		}
	}

	@PutMapping("{id}/cancel")
	public Map<String, Object> cancelReview(@PathVariable("id") long id) {
		if (!settingsService.is(ServerSetting.TASKS_ENABLED))
			throw Response.unavailable("Task feature not enabled");
		var review = service.get(id);
		if (review == null)
			throw Response.notFound("No review with id " + id + " found");
		try (var repo = repoService.get(review.repositoryPath)) {
			service.end(review, TaskState.CANCELED);
			notificationService.taskCanceled(repo, review).send();
			return getActiveTasks();
		}
	}

	@PutMapping("{id}/assign/{username}")
	public Map<String, Object> assignReviewer(
			@PathVariable("id") long id,
			@PathVariable("username") String username) {
		if (!settingsService.is(ServerSetting.TASKS_ENABLED))
			throw Response.unavailable("Task feature not enabled");
		var review = service.get(id);
		if (review == null)
			throw Response.notFound("No review with id " + id + " found");
		if (review.references.isEmpty())
			throw Response.badRequest("", "Please select data set references before assigning a user");
		try (var repo = repoService.get(review.repositoryPath)) {
			var assignment = service.startAssignment(review, username,
					(user, r) -> accessService.canReviewIn(user, r.path()));
			notificationService.taskAssigned(repo, review, assignment).send();
			return getActiveTasks();
		} catch (WebRequestException e) {
			throw Response.status(e);
		}
	}

	@PutMapping("{id}/complete/{username}")
	public Map<String, Object> completeAssignment(
			@PathVariable("id") long id,
			@PathVariable("username") String username) {
		if (!settingsService.is(ServerSetting.TASKS_ENABLED))
			throw Response.unavailable("Task feature not enabled");
		var review = service.get(id);
		if (review == null)
			throw Response.notFound("No review with id " + id + " found");
		try (var repo = repoService.get(review.repositoryPath)) {
			if (review.state != TaskState.PROCESSING)
				throw Response.conflict("Task is not in process state");
			var assignment = service.endAssignment(review, username, false);
			if (assignment == null)
				throw Response.notFound("User " + username + " has no active assignment");
			notificationService.taskCompleted(repo, review, assignment);
			return getActiveTasks();
		}
	}

	@PutMapping("{id}/cancel/{username}")
	public Map<String, Object> cancelAssignment(
			@PathVariable("id") long id,
			@PathVariable("username") String username) {
		if (!settingsService.is(ServerSetting.TASKS_ENABLED))
			throw Response.unavailable("Task feature not enabled");
		var review = service.get(id);
		if (review == null)
			throw Response.notFound("No review with id " + id + " found");
		try (var repo = repoService.get(review.repositoryPath)) {
			if (review.state != TaskState.PROCESSING)
				throw Response.conflict("Task is not in process state");
			var assignment = service.endAssignment(review, username, true);
			if (assignment == null)
				throw Response.notFound("User " + username + " has no active assignment");
			notificationService.taskRevoked(repo, review, assignment);
			return getActiveTasks();
		}
	}

	@PutMapping("{id}/markAsReviewed/{referenceId}/{value}")
	public Map<String, Object> markAsReviewed(
			@PathVariable("id") long id,
			@PathVariable("referenceId") long referenceId,
			@PathVariable("value") boolean value) {
		if (!settingsService.is(ServerSetting.TASKS_ENABLED))
			throw Response.unavailable("Task feature not enabled");
		var review = service.get(id);
		if (review == null)
			throw Response.notFound("No review with id " + id + " found");
		service.markAsReviewed(id, referenceId, value);
		return getActiveTasks();
	}

	private Map<String, Object> getActiveTasks() {
		// TODO return integer
		var user = userService.getCurrentUser();
		var activeTasks = taskService.getAllActiveFor(user).size();
		return Collections.singletonMap("activeTasks", Integer.toString(activeTasks));
	}

	private ReviewReference convert(Repository repo, Reference ref) {
		var reference = new ReviewReference();
		reference.type = ref.type;
		reference.refId = ref.refId;
		reference.commitId = ref.commitId;
		// TODO set name
		// reference.name = ref.name;
		return reference;
	}

}